gui/requirements.txt
# Phase 1 Core Communication Dependencies
pyserial>=3.5